Remove code that is merged into RL core
Remove cooking, dialogue, movement, music, woodcutting, thieving, ZMI. These are all in the core plugin.

This is being slightly refocused (for a little while - maybe a couple weeks) to get information about all transmitted varbits.
