+++ runelite-plugin.properties
@@ -3,4 +3,4 @@ author=OSRS Wiki
 support=https://osrs.wiki/RS:CROWD
 description=Help the wiki figure out skilling success rates, burn rates, and more.
 tags=wiki
-plugins=com.Crowdsourcing.CrowdsourcingPlugin
+plugins=com.Crowdsourcing.AdvancedCrowdsourcingPlugin
