--- src/main/java/com/Crowdsourcing/dialogue/DialogueOptionsData.java
@@ -1,10 +1,0 @@
-package com.Crowdsourcing.dialogue;
-
-import lombok.Data;
-import lombok.AllArgsConstructor;
-
-@Data
-@AllArgsConstructor
-public class DialogueOptionsData {
-	private final String[] options;
-}
