--- src/main/java/com/Crowdsourcing/skilling/SkillingEndReason.java
@@ -1,8 +1,0 @@
-package com.Crowdsourcing.skilling;
-
-public enum SkillingEndReason
-{
-	DEPLETED,
-	INVENTORY_FULL,
-	INTERRUPTED
-}
