--- src/main/java/com/Crowdsourcing/skilling/SkillingState.java
@@ -1,9 +1,0 @@
-package com.Crowdsourcing.skilling;
-
-public enum SkillingState
-{
-	READY,
-	CLICKED,
-	CUTTING,
-	RECOVERING
-}
